planning_cctp/requirements.txt
streamlit==1.39.0
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.3.0
spacy==3.7.4
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.7.0/fr_core_news_lg-3.7.0-py3-none-any.whl
pandas==2.2.2
openpyxl==3.1.2
networkx==3.2.1
